backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
nltk>=3.8.1
scikit-learn>=1.2.2
numpy>=1.24.3
python-dotenv>=1.0.0
django-cors-headers>=4.1.0
mssql-django>=1.3
pyodbc>=4.0.39 